+++ src/stream_pipe.h
@@ -4,6 +4,7 @@
 #if defined(NODE_WANT_INTERNALS) && NODE_WANT_INTERNALS
 
 #include "stream_base.h"
+#include "allocated_buffer.h"
 
 namespace node {
